requirements.txt
python-dotenv>=1.0.0
jinja2>=3.1.2
openpyxl>=3.1.2
requests>=2.31.0
